Improve RTConversionError
Added type (conversion type attempted) and value (the value that caused
the error) args to make it easier to provide info about conversion
errors. The detail arg from RTError remains to provide conversion-
specific detail.

Note that this matches the existing usage of RTConversionError when
a datetime conversion fails in RTDataSerializer.convert_to_datetime(),
so this is also a minor bug fix.