Include all detail lines in the detail attr. of RTResponse instances
Previously, we only included the first detail line, but there's no good
reason to not include them all. In most cases, there will only be one
detail line anyway.

Because of this, we now have to check all detail lines for the expected
text when doing RT API operations. This is probably better because we
always assumed the expected text would be in the first detail line, but
I'm not sure that would always be the case.